Move all structs and typedefs into a types.h file to simplify necessary includes (#32)
