+++ src/lexer.c
@@ -1,6 +1,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+
+#include "types.h"
 #include "util.h"
 #include "lexer.h"
