+++ src/sexpr.c
@@ -2,6 +2,7 @@
 #include <stdlib.h>
 #include <string.h>
 
+#include "types.h"
 #include "util.h"
 #include "env.h"
 #include "sexpr.h"
